Updated Element interface to use AttributeMap and NamespaceMap for xml_attributes and xmlns_attributes respectively. Element.xml_namespaces now also uses a NamespaceMap for its result to facilitate its use as the GetAllNs replacement.
--HG--
rename : amara/src/domlette/namednodemap.c => amara/src/domlette/attributemap.c
rename : amara/src/domlette/namednodemap.h => amara/src/domlette/attributemap.h

+++ amara/xpath/src/datatypes.c
@@ -346,8 +346,9 @@ static PyObject *Number_New(PyObject *value)
   }
   else if (value->ob_type->tp_as_buffer) {
     const char *buffer;
-    int len;
+    Py_ssize_t len;
     if (PyObject_AsCharBuffer(value, &buffer, &len)) return NULL;
+    /* FIXME: update FROM_STRING() to use `len` instead of assuming 0-term */
     FROM_STRING(char, buffer);
   }
